separate init module
